more stuff

+++ Assets/Utilities/FileWriter.cs
@@ -4,16 +4,13 @@ namespace Utilities
 {
     public class FileWriter
     {
-        private readonly System.IO.StreamWriter _file;
+        private readonly StreamWriter _file;
 
         public FileWriter(string directory, string filename)
         {
             // TODO: remove this and replace usage with a StreamWriter
-            string fullFilename = $"{directory}/{filename}.csv";
-            if (File.Exists(fullFilename))
-            {
-                File.Delete(fullFilename);
-            }
+            var fullFilename = $"{directory}/{filename}.csv";
+            if (File.Exists(fullFilename)) File.Delete(fullFilename);
             _file = new StreamWriter(fullFilename);
         }
 
@@ -21,7 +18,7 @@ namespace Utilities
         {
             _file.Write(text);
         }
-        
+
         public void WriteLine(string text)
         {
             _file.WriteLine(text);
